LIBCIR-23. Oracle JDK setup.
The bootstrap provisioning script looks for a JDK (without caring about version) in the /vagrant/dist directory. After installing on (if found), it looks for an /apps/jdk* directory, and uses that value to set the JAVA_HOME in the /etc/profile.d/mdsoar.sh config. Puppet no longer handles that configuration.

https://issues.umd.edu/browse/LIBCIR-23